Disambiguate other app in readme
I landed on this repo via Google looking for info about an [app in the Play store](https://play.google.com/store/apps/details?id=com.dev47apps.droidcam) which also happens to be called DroidCam.

Obviously naming collisions like this happen, I'm not asking you to renamed this or anything silly like that, but I did lose a couple of minutes before I realised this repo is unrelated to that app. This patch simply aims to reduce the scope for confusion with a (discreet but visible) line near the top of readme.md noting the difference.